cloud_llm_client: Move `CompletionIntent` to `language_model`
