+++ crates/agent/src/edit_agent.rs
@@ -8,7 +8,6 @@ pub mod streaming_fuzzy_matcher;
 use crate::{Template, Templates};
 use action_log::ActionLog;
 use anyhow::Result;
-use cloud_llm_client::CompletionIntent;
 use create_file_parser::{CreateFileParser, CreateFileParserEvent};
 pub use edit_parser::EditFormat;
 use edit_parser::{EditParser, EditParserEvent, EditParserMetrics};
@@ -21,8 +20,8 @@ use futures::{
 use gpui::{AppContext, AsyncApp, Entity, Task};
 use language::{Anchor, Buffer, BufferSnapshot, LineIndent, Point, TextBufferSnapshot};
 use language_model::{
-    LanguageModel, LanguageModelCompletionError, LanguageModelRequest, LanguageModelRequestMessage,
-    LanguageModelToolChoice, MessageContent, Role,
+    CompletionIntent, LanguageModel, LanguageModelCompletionError, LanguageModelRequest,
+    LanguageModelRequestMessage, LanguageModelToolChoice, MessageContent, Role,
 };
 use project::{AgentLocation, Project};
 use reindent::{IndentDelta, Reindenter};
